Unmock and disable tests in the React renderer
Summary:
We should unmock React just because we do that elsewhere and tests work better
that way. I was trying to get React tests working in this repo but because
we do so many special things in the React Core repo I gave up.

These test run in the React Core repo already.

Reviewed By: spicyj, bvaughn

Differential Revision: D4541126

fbshipit-source-id: ffbb1b76aac910a976222db91b80b8839fcecc60