Add UberEATS article to showcase
Summary:
Adds UberEATS technical blog post, https://eng.uber.com/ubereats-react-native/, to the showcase.

This is not an ideal entry, as the app in question, the UberEATS Restaurant Dashboard, is not widely available in the app stores for people to download and try out. I still think it's worth adding this to the showcase as a good example of a large, known company leveraging React Native to build a new experience.

Test Plan
=======

Built website and verified everything rendered correctly in the Showcase.
Closes https://github.com/facebook/react-native/pull/13274

Differential Revision: D4823507

Pulled By: ericvicenti

fbshipit-source-id: 7929124dfc17254d1bc5970a20ba393a54e34409